+++ felix/dataplane/linux/bpf_ep_mgr.go
@@ -2370,7 +2370,7 @@ func (m *bpfEndpointManager) doApplyPolicy(ifaceName string) (bpfInterfaceState,
 
 	ap.DSCP = -1
 	if wep != nil && len(wep.QosPolicies) > 0 {
-		ap.DSCP = int16(wep.QosPolicies[0].Dscp)
+		ap.DSCP = int8(wep.QosPolicies[0].Dscp)
 	}
 
 	if err := m.wepStateFillJumps(ap, &state); err != nil {
@@ -2794,7 +2794,8 @@ func (d *bpfEndpointManagerDataplane) attachDataIfaceProgram(
 
 	ap.DSCP = -1
 	if ep != nil && len(ep.QosPolicies) > 0 {
-		ap.DSCP = int16(ep.QosPolicies[0].Dscp)
+		// TODO (mazdak): panic?
+		ap.DSCP = int8(ep.QosPolicies[0].Dscp)
 	}
 
 	if ep != nil {
